requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
mcp>=1.0.0
uvicorn>=0.32.1
PySide6>=6.5.0

openai>=1.1.0
anthropic>=0.49.0

openai
docutils
beautifulsoup4
Pygments
langchain
langchain-community
langchain-openai
faiss-cpu
tiktoken